% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_connect}
\alias{db_connect}
\title{function for connectiong to db}
\usage{
db_connect(path = "")
}
\arguments{
\item{path}{path to database}
}
\description{
function for connectiong to db
}

