% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_disconnect}
\alias{db_disconnect}
\title{function for disconnecting from db}
\usage{
db_disconnect(db)
}
\arguments{
\item{db}{connection to db}
}
\description{
function for disconnecting from db
}

