% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_ensure_exists_dishes}
\alias{db_ensure_exists_dishes}
\title{function for ensuring that dishes table exists in db}
\usage{
db_ensure_exists_dishes(path = "")
}
\description{
function for ensuring that dishes table exists in db
}

