% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_ensure_exists_tweets}
\alias{db_ensure_exists_tweets}
\title{function for ensuring that tweets table exists in db}
\usage{
db_ensure_exists_tweets(path = "")
}
\description{
function for ensuring that tweets table exists in db
}

