% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_ensure_table_exists}
\alias{db_ensure_table_exists}
\title{function for ensuring that a particular table exists in db}
\usage{
db_ensure_table_exists(table = "", path = "")
}
\description{
function for ensuring that a particular table exists in db
}

