% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_dish_data}
\alias{db_get_dish_data}
\title{function for retrieving data from dishes table in db}
\usage{
db_get_dish_data(date = Sys.Date(), loc = "mensa_giessberg", lang = "de")
}
\description{
function for retrieving data from dishes table in db
}

