% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_request_data}
\alias{db_get_request_data}
\title{function for retrieving data from requests table in db}
\usage{
db_get_request_data(date = Sys.Date(), status = 200,
  loc = "mensa_giessberg", lang = "de")
}
\description{
function for retrieving data from requests table in db
}

