% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_path}
\alias{db_path}
\title{function for determining the path to the db used for storage}
\usage{
db_path(path = "")
}
\arguments{
\item{path}{path to database}
}
\description{
function for determining the path to the db used for storage
}

