% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweeting.R
\docType{data}
\name{dings}
\alias{dings}
\title{function for generating tweets}
\format{\preformatted{ num 1
}}
\usage{
dings
}
\description{
function for generating tweets
}
\keyword{datasets}

