% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{even}
\alias{even}
\title{function giving back elements with even index}
\usage{
even(x)
}
\arguments{
\item{x}{for which to give back odd index}
}
\description{
function giving back elements with even index
}

