% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweeting.R
\name{gen_tweets}
\alias{gen_tweets}
\title{function for generating tweets}
\usage{
gen_tweets(date = Sys.Date(), lang = "de", loc = "mensa_giessberg",
  type = "all", regex = "^.*", invert = FALSE)
}
\arguments{
\item{date}{date for which tweets shall be generated}

\item{lang}{language of tweets}

\item{loc}{location of mensa}

\item{regex}{possible regex for further subsetting}

\item{invert}{shall regex be inverted?}

\item{dish}{types}
}
\description{
function for generating tweets
}

