% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_additives}
\alias{get_additives}
\title{function translating additive numbers into description}
\usage{
get_additives(x)
}
\arguments{
\item{x}{vector of additives}
}
\description{
function translating additive numbers into description
}

