% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{httr_content_as_list}
\alias{httr_content_as_list}
\title{function for making lists with prefixed names}
\usage{
httr_content_as_list(x, prefix = "")
}
\arguments{
\item{x}{named vector to make list of}
}
\description{
function for making lists with prefixed names
}

