% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{mensaplan}
\alias{mensaplan}
\title{wrapper function for get_mensaplan() and parse_mensaplan}
\usage{
mensaplan(date = format(Sys.Date(), "\%Y-\%m-\%d"), lang = "de",
  loc = "mensa_giessberg")
}
\arguments{
\item{date}{date for which to get data}

\item{lang}{language of the mensa plan -- either "de" or "en"}

\item{loc}{location identifier}
}
\description{
wrapper function for get_mensaplan() and parse_mensaplan
}

