% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{mp_data}
\alias{mp_data}
\title{function for reading the contents of the local mensaplan database}
\usage{
mp_data(path = "")
}
\description{
function for reading the contents of the local mensaplan database
}

