% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{mp_data_retrieval}
\alias{mp_data_retrieval}
\title{function for managing data retrieval
downloads data (if necessary) and stores requests in database}
\usage{
mp_data_retrieval(lang = c("de", "en"), date = format(Sys.Date(),
  "\%Y-\%m-\%d"), loc = "mensa_giessberg", force = FALSE)
}
\arguments{
\item{lang}{paramenter for HTTP POST determining the language of data}

\item{date}{paramenter for HTTP POST determining the date of the dish}

\item{loc}{paramenter for HTTP POST determining the location for which to
retrieve data}

\item{force}{by default the function will not download data for which it got
an valid (HTTP 200) response already - if set to true it will do it
anyways}
}
\description{
function for managing data retrieval
downloads data (if necessary) and stores requests in database
}

