% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{mp_http_post}
\alias{mp_http_post}
\title{function POSTing HTTP requests to data server}
\usage{
mp_http_post(lang = c("de", "en"), date = format(Sys.Date(), "\%Y-\%m-\%d"),
  loc = "mensa_giessberg")
}
\arguments{
\item{lang}{language of the mensa plan -- either "de" or "en"}

\item{date}{date for which to get data}

\item{loc}{location identifier}
}
\description{
function POSTing HTTP requests to data server
}

