% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mensaplan.R
\name{mp_parse}
\alias{mp_parse}
\title{function parsing data retrieved}
\usage{
mp_parse(post_results)
}
\arguments{
\item{post_results}{results from get_mesaplan()}
}
\description{
function parsing data retrieved
}

