% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{mp_save}
\alias{mp_save}
\title{function for saving mensaplan to disk}
\usage{
mp_save(res, path = "")
}
\arguments{
\item{res}{the result of a call to mensaplan() or parse_mensaplan()}
}
\description{
function for saving mensaplan to disk
}

