% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mensaplan.R
\name{mp_scrape}
\alias{mp_scrape}
\title{function POSTing HTTP requests to data server}
\usage{
mp_scrape(lang = c("de", "en"), date = format(Sys.Date(), "\%Y-\%m-\%d"),
  loc = "mensa_giessberg")
}
\arguments{
\item{lang}{language of the mensa plan -- either "de" or "en"}

\item{date}{date for which to get data}

\item{loc}{location identifier}
}
\description{
function POSTing HTTP requests to data server
}

