% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{odd}
\alias{odd}
\title{function giving back elements with odd index}
\usage{
odd(x)
}
\arguments{
\item{x}{vector for which to give back even index}
}
\description{
function giving back elements with odd index
}

