% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mensaplan.R
\name{print.mensaplan}
\alias{print.mensaplan}
\title{customized print function for mensaplan data.frames}
\usage{
\method{print}{mensaplan}(x, ..., digits = NULL, quote = FALSE,
  right = TRUE, row.names = TRUE)
}
\description{
customized print function for mensaplan data.frames
}

