% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{request_to_dish}
\alias{request_to_dish}
\title{function parsing data retrieved}
\usage{
request_to_dish(res)
}
\arguments{
\item{post_results}{results from get_mesaplan()}
}
\description{
function parsing data retrieved
}

