% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\docType{data}
\name{storage}
\alias{storage}
\title{storage for unikonstanzmensabot infos}
\format{An object of class \code{environment} of length 3.}
\usage{
storage
}
\description{
storage for unikonstanzmensabot infos
}
\keyword{datasets}

