% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_connect}
\alias{db_connect}
\title{function for connectiong to db used for storage}
\usage{
db_connect(path = "")
}
\arguments{
\item{path}{path to database}
}
\description{
function for connectiong to db used for storage
}

