% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mensaplan.R
\name{mensaplan}
\alias{mensaplan}
\title{wrapper function for get_mensaplan() and parse_mensaplan}
\usage{
mensaplan(lang = c("de", "en"), date = format(Sys.Date(), "\%Y-\%m-\%d"),
  loc = "mensa_giessberg")
}
\arguments{
\item{lang}{language of the mensa plan -- either "de" or "en"}

\item{date}{date for which to get data}

\item{loc}{location identifier}
}
\description{
wrapper function for get_mensaplan() and parse_mensaplan
}

